==> 00_install_pkgs.R <==
install.packages(c("dplyr","GGally","BiocManager","matrixStats","ggplot2","data.table","tibble","tidyverse","janitor","gridExtra","circlize","magick"), repos="https://cloud.r-project.org")
BiocManager::install(c("ComplexHeatmap","limma","swamp"),force=T)
